Queue all device commands and execute them from the main thread.
By queueing all commands the client library will return immediately and all
executed commands will not block any more. This is benificial for applications
that calls the api using the same thread as the ui thread. For instance groups
could before take several seconds to execute and the application would freeze in
the meantime. If the time to execute was longer than the timeout between
telldus-core and telldusd other nasty stuff like resending the same group up to
20 times could happen.

This change fixes all that with the drawback that we do not really know if the
command was succuessful when returning. We do the best to check all
prerequisites to determine it the call will be successful. Errors like
TELLSTICK_ERROR_COMMUNICATION and TELLSTICK_ERROR_BROKEN_PIPE will not be
reported back to the client. We can still log them though.

This fixes #260. Please reopen if the issue still exists.